homeassistant/scripts/requirements.txt
i2cdevice 
smbus
RPi.bme280
